requirements2.txt
aniso8601==9.0.1
boto3==1.20.27
botocore==1.23.27
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.2
protobuf==3.19.1
pyarrow==5.0.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
pyxirr==0.7.1
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.1
